+++ src/components/Main/Reports/Reports.styles.jsx
@@ -1,0 +1,15 @@
+import { styled } from "styled-components";
+
+export const $Reports = styled.div`
+  display: block;
+
+  & > * {
+    margin-bottom: 15px;
+  }
+`;
+
+export const $Fallback = styled.div`
+  display: flex;
+  justify-content: center;
+  margin: 27px 0 28px 0;
+`;