+++ src/components/Main/Reports/ReportsForm/Button/Button.jsx
@@ -1,0 +1,11 @@
+import { $Button } from "./Button.styles.jsx";
+
+export default function Button({ enabled, name, children, ...props}) {
+
+    return(
+        <$Button $enabled={enabled} {...props}>
+            <div>{ children }</div>
+            &nbsp;{ name }
+        </ $Button>
+    )
+}