+++ src/components/Main/Reports/ReportsForm/Label/Label.jsx
@@ -1,0 +1,10 @@
+import { $Label } from "./Label.styles.jsx";
+
+export default function BuilderButton({ name, children }) {
+  return (
+    <$Label>
+      <div>{children}</div>
+      &nbsp;{name}
+    </$Label>
+  );
+}